backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
python-multipart==0.0.6
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
aiofiles==23.2.1
brotli==1.1.0
Pillow>=9.0.0
python-dotenv>=1.0.1
openai>=1.40.2
openpyxl>=3.1.2
sse-starlette>=1.6.1
redis>=5.0.8
